google-cloud-bigtable-try/pyproject.toml
[project]
name = "google-cloud-bigtable-try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigtable==2.35.0",
    "google-cloud-core==2.5.0",
]
